+++ src/goals/ReviewEntries/ReviewEntriesComponent/ReviewEntriesReducer.tsx
@@ -7,15 +7,11 @@ import {
 export interface ReviewEntriesState {
   words: ReviewEntriesWord[];
   language: string;
-  isRecording: boolean;
-  wordBeingRecorded: string | undefined;
 }
 
 export const defaultState: ReviewEntriesState = {
   words: [],
   language: "en",
-  isRecording: false,
-  wordBeingRecorded: undefined,
 };
 
 export const reviewEntriesReducer = (
@@ -47,14 +43,6 @@ export const reviewEntriesReducer = (
             };
           } else return word;
         }),
-      };
-
-    case ReviewEntriesActionTypes.UpdateRecordingStatus:
-      // Update recording status
-      return {
-        ...state,
-        isRecording: action.recordingStatus,
-        wordBeingRecorded: action.wordId,
       };
 
     default:
